enhance shell detection logic and add tests
