+++ internal/telemetry/properties.go
@@ -18,16 +18,13 @@ import (
 	"context"
 	"crypto/rand"
 	"encoding/hex"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"runtime"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/datarobot/cli/internal/config"
 	"github.com/datarobot/cli/internal/config/viperx"
+	"github.com/datarobot/cli/internal/shell"
 	"github.com/datarobot/cli/internal/version"
 )
 
@@ -58,61 +55,16 @@ type CommonProperties struct {
 	CommandKind       string // "core" or "plugin", set by the root command after dispatch
 }
 
-// parentProcessName returns the name of the process that launched the CLI
-// (i.e. the running shell). On Linux it reads /proc/{ppid}/comm; on
-// macOS and other Unix systems it falls back to querying ps. Returns an
-// empty string when the name cannot be determined.
-func parentProcessName() string {
-	ppid := os.Getppid()
-
-	// Linux: /proc/{ppid}/comm contains the short process name
-	if data, err := os.ReadFile("/proc/" + strconv.Itoa(ppid) + "/comm"); err == nil {
-		return strings.TrimSpace(string(data))
-	}
-
-	// macOS and other Unix: ask ps for the command name
-	out, err := exec.Command("ps", "-p", strconv.Itoa(ppid), "-o", "comm=").Output()
-	if err != nil {
-		return ""
-	}
-
-	name := strings.TrimSpace(string(out))
-
-	return filepath.Base(name)
-}
-
-// detectShellFromEnv falls back to environment variables when the parent
-// process name cannot be determined. $SHELL reflects the login shell and
-// may differ from the shell that is currently running the CLI.
-func detectShellFromEnv() string {
-	// Unix/macOS: $SHELL holds the path to the login shell (e.g. "/bin/zsh")
-	if shell := os.Getenv("SHELL"); shell != "" {
-		return filepath.Base(shell)
-	}
-
-	// Windows PowerShell sets $PSModulePath
-	if os.Getenv("PSModulePath") != "" {
-		return "powershell"
-	}
-
-	// Windows cmd.exe sets $ComSpec
-	if os.Getenv("ComSpec") != "" {
-		return "cmd"
-	}
-
-	return "unknown"
-}
-
 // DetectShell returns the name of the shell the CLI is running from.
-// It first inspects the parent process (accurate even after exec sh, exec bash,
-// etc.) and falls back to $SHELL / Windows env vars when that is unavailable.
-// Returns "unknown" if the shell cannot be determined.
+// Delegates to shell.DetectShell() which inspects the parent process first
+// (accurate even after exec sh/bash/etc.) and falls back to $SHELL / OS
+// defaults. Returns "unknown" if the shell cannot be determined.
 func DetectShell() string {
-	if name := parentProcessName(); name != "" {
+	if name, err := shell.DetectShell(); err == nil {
 		return name
 	}
 
-	return detectShellFromEnv()
+	return "unknown"
 }
 
 // CollectCommonProperties gathers all common telemetry properties from the
